refactor: make `PartitionKeyNames` a top-level class
It's not really a MicroMeter utility, it's just where
we define common labels for everything partition-scoped.

We can't move this type to the broker module because it's
required in various leaf modules such as snapshots.
