+++ zeebe/broker/src/main/java/io/camunda/zeebe/broker/system/monitoring/HealthMetricsDoc.java
@@ -8,7 +8,7 @@
 package io.camunda.zeebe.broker.system.monitoring;
 
 import io.camunda.zeebe.util.micrometer.ExtendedMeterDocumentation;
-import io.camunda.zeebe.util.micrometer.MicrometerUtil.PartitionKeyNames;
+import io.camunda.zeebe.util.micrometer.PartitionKeyNames;
 import io.micrometer.common.docs.KeyName;
 import io.micrometer.core.instrument.Meter;
