+++ zeebe/journal/src/main/java/io/camunda/zeebe/journal/file/JournalMetricsDoc.java
@@ -8,7 +8,7 @@
 package io.camunda.zeebe.journal.file;
 
 import io.camunda.zeebe.util.micrometer.ExtendedMeterDocumentation;
-import io.camunda.zeebe.util.micrometer.MicrometerUtil.PartitionKeyNames;
+import io.camunda.zeebe.util.micrometer.PartitionKeyNames;
 import io.micrometer.common.docs.KeyName;
 import io.micrometer.core.instrument.Meter.Type;
 import java.time.Duration;
