+++ zeebe/logstreams/src/main/java/io/camunda/zeebe/logstreams/impl/log/SequencerMetrics.java
@@ -15,7 +15,7 @@ import static io.camunda.zeebe.logstreams.impl.log.SequencerMetrics.SequencerMet
 import io.camunda.zeebe.logstreams.impl.LogStreamMetricsDoc.FlowControlContext;
 import io.camunda.zeebe.logstreams.log.WriteContext;
 import io.camunda.zeebe.util.micrometer.ExtendedMeterDocumentation;
-import io.camunda.zeebe.util.micrometer.MicrometerUtil.PartitionKeyNames;
+import io.camunda.zeebe.util.micrometer.PartitionKeyNames;
 import io.micrometer.common.docs.KeyName;
 import io.micrometer.core.instrument.DistributionSummary;
 import io.micrometer.core.instrument.Meter.Type;
