+++ zeebe/zb-db/src/main/java/io/camunda/zeebe/db/ColumnFamilyMetricsDoc.java
@@ -9,7 +9,7 @@ package io.camunda.zeebe.db;
 
 import io.camunda.zeebe.util.micrometer.ExtendedMeterDocumentation;
 import io.camunda.zeebe.util.micrometer.MicrometerUtil;
-import io.camunda.zeebe.util.micrometer.MicrometerUtil.PartitionKeyNames;
+import io.camunda.zeebe.util.micrometer.PartitionKeyNames;
 import io.micrometer.common.docs.KeyName;
 import io.micrometer.core.instrument.Meter.Type;
 import java.time.Duration;
